HBASE-13661 Correct binary compatibility issues discovered in 1.1.0RC0

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/MetricsMaster.java
@@ -52,7 +52,7 @@ public class MetricsMaster {
   /**
    * @param inc How much to add to requests.
    */
-  public void incrementRequests(final int inc) {
+  public void incrementRequests(final long inc) {
     masterSource.incRequests(inc);
   }
 }
